CommandWords showAllCommands return type changed to String
Thinking Ahead if later on the game changes its approach from a text based game to any graphical interface based game then printLn method will not be used. Leaving the choice to games class to show it to the user, I changed the Parser method to return a string and Game class to have a choice to display it however it likes.